feat: 파일 업로드 후 uri 권한 해제 로직 수정
