--- core/common-ui/src/main/java/com/speech/common_ui/ui/UploadFileDialog.kt
@@ -1,2 +1,0 @@
-package com.speech.common_ui.ui
-
