+++ core/common-ui/src/main/java/com/speech/common_ui/util/MediaUtil.kt
@@ -1,13 +1,16 @@
 package com.speech.common_ui.util
 
+import android.content.ContentResolver
 import android.content.Context
 import android.media.MediaMetadataRetriever
 import android.net.Uri
 import android.util.Log
+import android.webkit.MimeTypeMap
 import androidx.annotation.WorkerThread
 import com.speech.domain.model.speech.SpeechFileRule.MAX_DURATION_MS
 import com.speech.domain.model.speech.SpeechFileRule.MIN_DURATION_MS
 import com.speech.domain.model.speech.SpeechFileType
+import java.io.File
 
 object MediaUtil {
     @WorkerThread
