+++ core/data/src/main/java/com/speech/data/util/MediaUtil.kt
@@ -1,10 +1,9 @@
 package com.speech.data.util
 
 import android.content.ContentResolver
-import android.content.Context
-import android.media.MediaMetadataRetriever
 import android.net.Uri
 import android.webkit.MimeTypeMap
+import java.io.File
 
 fun getMimeType(contentResolver: ContentResolver, uri: Uri): String {
     return contentResolver.getType(uri) ?: throw IllegalArgumentException("Unknown URI")
@@ -19,3 +18,11 @@ fun getExtension(contentResolver: ContentResolver, uri: Uri): String {
     } ?: ""
 }
 
+fun getMimeType(file : File) : String {
+    val fileExtension = file.extension
+    val mimeType = MimeTypeMap.getSingleton().getMimeTypeFromExtension(fileExtension) ?: throw IllegalArgumentException("Unknown File")
+    return when (mimeType) {
+        "audio/x-wav" -> "audio/wav"
+        else -> mimeType
+    }
+}
